Manager: move next_nav_focus to public API; tweaks to menus
