+++ src/widget/menu/menubar.rs
@@ -110,7 +110,7 @@ impl<D: Directional, W: Widget<Msg = M>, M> event::Handler for MenuBar<D, W> {
                         && mgr.request_grab(self.id(), source, coord, GrabMode::Grab, None)
                     {
                         mgr.set_grab_depress(source, Some(start_id));
-                        mgr.set_nav_focus(Some(start_id));
+                        self.find(start_id).map(|w| mgr.next_nav_focus(w, false));
                         self.opening = false;
                         if self.rect().contains(coord) {
                             // We could just send Event::OpenPopup, but we also
@@ -140,15 +140,17 @@ impl<D: Directional, W: Widget<Msg = M>, M> event::Handler for MenuBar<D, W> {
                 }
             }
             Event::PressMove { source, cur_id, .. } => {
-                if cur_id.map(|id| self.is_ancestor_of(id)).unwrap_or(false) {
-                    let id = cur_id.unwrap();
-                    mgr.set_grab_depress(source, Some(id));
-                    mgr.set_nav_focus(Some(id));
-                    self.delayed_open = Some(id);
-                    mgr.update_on_timer(Duration::from_millis(300), self.id());
+                if let Some(w) = cur_id.and_then(|id| self.find(id)) {
+                    if w.key_nav() {
+                        let id = cur_id.unwrap();
+                        mgr.set_grab_depress(source, Some(id));
+                        mgr.set_nav_focus(id);
+                        self.delayed_open = Some(id);
+                        mgr.update_on_timer(Duration::from_millis(300), self.id());
+                    }
                 } else {
                     mgr.set_grab_depress(source, None);
-                    mgr.set_nav_focus(None);
+                    mgr.clear_nav_focus();
                 }
             }
             Event::PressEnd { coord, end_id, .. } => {
